Enhance NetworkStorage and SyncManager with collection-based support and synchronization features. Introduce NetworkStorageCollectionBasedConfig for configuration options, and implement methods for querying, updating, and deleting documents in a collection format. Update SyncManager to sync local storage with network data on fetch.

+++ lib/src/network_storage.dart
@@ -5,25 +5,62 @@ import 'package:flutter/services.dart';
 import 'std_obj.dart';
 import 'synced.dart';
 
+typedef QueryFn<T> = Query<T> Function(Query<T> colRef);
+
+class NetworkStorageCollectionBasedConfig<T> {
+  final bool getAllEnabled;
+  final int maxGetAllDocs;
+  final QueryFn<T>? defaultQuery;
+
+  const NetworkStorageCollectionBasedConfig({
+    this.getAllEnabled = true,
+    this.maxGetAllDocs = 10,
+    this.defaultQuery,
+  });
+}
+
 class NetworkStorage<T> {
-  final String defaultDocPath;
+  final String path;
   final bool disabled;
 
+  /// If true, the data will be stored in a collection-based format,
+  /// i.e for each record there will be a separate document.
+  ///
+  /// NOTE:- This might lead to higher read/write costs in Firestore.
+  final bool collectionBased;
+  final NetworkStorageCollectionBasedConfig<T>? collectionBasedConfig;
+
   const NetworkStorage.disabled()
-      : defaultDocPath = '',
-        disabled = true;
+      : path = '',
+        disabled = true,
+        collectionBased = false,
+        collectionBasedConfig = null;
 
   const NetworkStorage(
-    this.defaultDocPath, {
+    this.path, {
     this.disabled = false,
+    this.collectionBased = false,
+    this.collectionBasedConfig,
   });
 
   Future<IMap<String, T>> getAll(StdObjParams<T> params,
       [String? docPath]) async {
     _assertDisabled();
+    if (collectionBased) {
+      if (!(collectionBasedConfig?.getAllEnabled ?? false)) {
+        throw PlatformException(
+          code: 'GET_ALL_DISABLED',
+          message:
+              'NetworkStorage.getAll called for collection-based storage, but getAll is disabled in config.',
+          details: 'Path: $path',
+        );
+      }
+
+      return getQuery(params);
+    }
 
     try {
-      final docRef = FirebaseFirestore.instance.doc(docPath ?? defaultDocPath);
+      final docRef = FirebaseFirestore.instance.doc(docPath ?? path);
       return (await docRef.get())
           .data()!
           .map((key, value) => MapEntry(key, params.fromJson(value)))
@@ -36,9 +73,57 @@ class NetworkStorage<T> {
     }
   }
 
+  /// Only enabled for collection-based storage.
+  Future<IMap<String, T>> getQuery(StdObjParams<T> params,
+      [QueryFn<T>? query]) {
+    _assertDisabled();
+    if (!collectionBased) {
+      throw PlatformException(
+        code: 'COLLECTION_BASED_STORAGE_DISABLED',
+        message: 'Querying is only supported for collection-based storage.',
+      );
+    }
+
+    if (query == null && collectionBasedConfig?.defaultQuery == null) {
+      throw PlatformException(
+        code: 'NO_QUERY_PROVIDED',
+        message: 'No query function provided for collection-based storage.',
+        details:
+            'Currently we do not allow, by default, fetching all documents '
+            'from collection-based storage to avoid high read costs. '
+            'Please provide a query function to limit the number of documents '
+            'fetched.',
+      );
+    }
+
+    Query<T> colRef =
+        FirebaseFirestore.instance.collection(path).withConverter<T>(
+              fromFirestore: (snap, _) => params.fromJson(snap.data()!),
+              toFirestore: (obj, _) => params.toJson(obj),
+            );
+
+    colRef = (query ?? collectionBasedConfig!.defaultQuery!)(colRef);
+    if (collectionBasedConfig!.maxGetAllDocs > 0) {
+      colRef = colRef.limit(collectionBasedConfig!.maxGetAllDocs);
+    }
+
+    return colRef.get().then((querySnapshot) {
+      final dataMap = {
+        for (final doc in querySnapshot.docs) doc.id: doc.data(),
+      };
+      return dataMap.toIMap();
+    });
+  }
+
   Future<void> update(Dataset<T> data, StdObjParams<T> params,
       [String? docPath]) async {
     if (disabled) return;
+
+    if (collectionBased) {
+      final batch = FirebaseFirestore.instance.batch();
+      writeBatchUpdate(batch, data, params, docPath);
+      return batch.commit();
+    }
 
     final (docRef, setData, setOptions) =
         _getUpdateParams(data, params, docPath);
@@ -50,6 +135,16 @@ class NetworkStorage<T> {
       [String? docPath]) {
     if (disabled) return transaction;
 
+    if (collectionBased) {
+      final colRef = FirebaseFirestore.instance.collection(docPath ?? path);
+      for (final entry in data.entries) {
+        final docRef = colRef.doc(entry.key);
+        transaction.set(
+            docRef, params.toJson(entry.value), SetOptions(merge: true));
+      }
+      return transaction;
+    }
+
     final (docRef, setData, setOptions) =
         _getUpdateParams(data, params, docPath);
     return transaction.set(docRef, setData, setOptions);
@@ -59,6 +154,15 @@ class NetworkStorage<T> {
       WriteBatch batch, Dataset<T> data, StdObjParams<T> params,
       [String? docPath]) {
     if (disabled) return;
+
+    if (collectionBased) {
+      final colRef = FirebaseFirestore.instance.collection(docPath ?? path);
+      for (final entry in data.entries) {
+        final docRef = colRef.doc(entry.key);
+        batch.set(docRef, params.toJson(entry.value), SetOptions(merge: true));
+      }
+      return;
+    }
 
     final (docRef, setData, setOptions) =
         _getUpdateParams(data, params, docPath);
@@ -73,7 +177,7 @@ class NetworkStorage<T> {
       [String? docPath]) {
     _assertDisabled();
 
-    final docRef = FirebaseFirestore.instance.doc(docPath ?? defaultDocPath);
+    final docRef = FirebaseFirestore.instance.doc(docPath ?? path);
     final setData =
         data.unlock.map((key, value) => MapEntry(key, params.toJson(value)));
     final setOptions = SetOptions(merge: true);
@@ -84,6 +188,12 @@ class NetworkStorage<T> {
       [String? docPath]) async {
     if (disabled) return;
 
+    if (collectionBased) {
+      final batch = FirebaseFirestore.instance.batch();
+      writeBatchDelete(batch, data, params, docPath);
+      return batch.commit();
+    }
+
     final (docRef, deleteData) = _getDeleteParams(data, params, docPath);
     return docRef.update(deleteData);
   }
@@ -92,6 +202,14 @@ class NetworkStorage<T> {
       Transaction transaction, Dataset<T> data, StdObjParams<T> params,
       [String? docPath]) {
     if (disabled) return transaction;
+
+    if (collectionBased) {
+      final colRef = FirebaseFirestore.instance.collection(docPath ?? path);
+      for (final id in data.keys) {
+        transaction.delete(colRef.doc(id));
+      }
+      return transaction;
+    }
 
     final (docRef, deleteData) = _getDeleteParams(data, params, docPath);
     return transaction.update(docRef, deleteData);
@@ -102,8 +220,16 @@ class NetworkStorage<T> {
       [String? docPath]) {
     if (disabled) return;
 
+    if (collectionBased) {
+      final colRef = FirebaseFirestore.instance.collection(docPath ?? path);
+      for (final id in data.keys) {
+        batch.delete(colRef.doc(id));
+      }
+      return;
+    }
+
     final (docRef, deleteData) = _getDeleteParams(data, params, docPath);
-    return batch.update(docRef, deleteData);
+    batch.update(docRef, deleteData);
   }
 
   (DocumentReference<Map<String, dynamic>>, Map<String, FieldValue>)
@@ -112,7 +238,7 @@ class NetworkStorage<T> {
     _assertDisabled();
 
     final ids = data.keys;
-    final docRef = FirebaseFirestore.instance.doc(docPath ?? defaultDocPath);
+    final docRef = FirebaseFirestore.instance.doc(docPath ?? path);
     final deleteData = {
       for (final id in ids) id: FieldValue.delete(),
     };
@@ -122,7 +248,19 @@ class NetworkStorage<T> {
   Future<void> clear([String? docPath]) async {
     if (disabled) return;
 
-    final docRef = FirebaseFirestore.instance.doc(docPath ?? defaultDocPath);
+    if (collectionBased) {
+      final colRef = FirebaseFirestore.instance.collection(docPath ?? path);
+      final snapshot = await colRef.get();
+      if (snapshot.docs.isEmpty) return;
+
+      final batch = FirebaseFirestore.instance.batch();
+      for (final doc in snapshot.docs) {
+        batch.delete(doc.reference);
+      }
+      return batch.commit();
+    }
+
+    final docRef = FirebaseFirestore.instance.doc(docPath ?? path);
     return docRef.delete();
   }
